refactor: challenge update
